feat: move deref implementation to QuestionnaireService
